subtensor_custom_rpc

Trait SubtensorCustomApiServer

pub trait SubtensorCustomApiServer<BlockHash>:
    Sized
    + Send
    + Sync
    + 'static {

    // Required methods
    fn get_delegates(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>;
    fn get_delegate(
        &self,
        delegate_account_vec: Vec<u8>,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_delegated(
        &self,
        delegatee_account_vec: Vec<u8>,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_neurons_lite(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_neuron_lite(
        &self,
        netuid: u16,
        uid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_neurons(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_neuron(
        &self,
        netuid: u16,
        uid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_subnet_info(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_subnets_info(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>;
    fn get_subnet_info_v2(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_subnets_info_v2(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>;
    fn get_subnet_hyperparams(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_all_dynamic_info(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>;
    fn get_dynamic_info(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_all_metagraphs(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>;
    fn get_metagraph(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_subnet_state(
        &self,
        netuid: u16,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;
    fn get_network_lock_cost(&self, at: Option<BlockHash>) -> RpcResult<u64>;
    fn get_selective_metagraph(
        &self,
        netuid: u16,
        metagraph_index: Vec<u16>,
        at: Option<BlockHash>,
    ) -> RpcResult<Vec<u8>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where BlockHash: Send + Sync + 'static + DeserializeOwned { ... }

}

Server trait implementation for the SubtensorCustomApi RPC API.

Required Methods

fn get_delegates(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_delegate( &self, delegate_account_vec: Vec<u8>, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_delegated( &self, delegatee_account_vec: Vec<u8>, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_neurons_lite( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_neuron_lite( &self, netuid: u16, uid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_neurons(&self, netuid: u16, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_neuron( &self, netuid: u16, uid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_subnet_info( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_subnets_info(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_subnet_info_v2( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_subnets_info_v2(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_subnet_hyperparams( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_all_dynamic_info(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_dynamic_info( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_all_metagraphs(&self, at: Option<BlockHash>) -> RpcResult<Vec<u8>>

fn get_metagraph( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_subnet_state( &self, netuid: u16, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

fn get_network_lock_cost(&self, at: Option<BlockHash>) -> RpcResult<u64>

fn get_selective_metagraph( &self, netuid: u16, metagraph_index: Vec<u16>, at: Option<BlockHash>, ) -> RpcResult<Vec<u8>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>
where BlockHash: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.